Fix `is_pydantic_dataclass` check condition to work with dataclasses in the process of being built
Do not rely on `__pydantic_validator__`, an attribute which is set
only when the class is fully built. Because we check for this condition
during schema building, we end up considering the class as a stdlib
dataclass.

As a consequence, the deepcopy logic of `FieldInfo` instances was
tweaked, as in some tests, the deep copy would fail. Instead, make
a shallow copy of every `FieldInfo` instance, this is enough
(the same thing is done in `collect_model_fields` for Pydantic
models, we make a shallow copy of the parent fields).